backend/requirements.txt
fastapi==0.115.0
uvicorn>=0.34.0
requests==2.32.3
python-dotenv==1.0.1
pydantic==2.9.2
httpx>=0.27.0,<0.29.0
google-adk==0.1.0
google-auth-oauthlib==1.2.1
google-api-python-client>=2.157.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
alembic==1.13.3
